requirements.txt
facenet-pytorch==2.6.0
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.2.0
qdrant-client==1.11.2
torch==2.2.2
